+++ Gestion_cadena_restaurante/iteracion1/src/main/java/org/iteracion1/Dominio/Estados.java
@@ -1,0 +1,12 @@
+package Dominio;
+
+
+public enum Estados { Libre, 
+	Reservada, 
+	Ocupada, 
+	pidiendo, 
+	espera_comida, 
+	servida, 
+	espera_cuenta, 
+	pagando, 
+	preparacion}
