+++ Gestion_cadena_restaurante/iteracion1/src/main/java/org/iteracion1/Dominio/Jefe_de_Sala.java
@@ -1,0 +1,25 @@
+package Dominio;
+
+
+public class Jefe_de_Sala {
+
+    private int id_jefe_sala;
+    private String nombre;
+
+    public int getId_jefe_sala() {
+        return id_jefe_sala;
+    }
+
+    public void setId_jefe_sala(int id_jefe_sala) {
+        this.id_jefe_sala = id_jefe_sala;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+}
