+++ Gestion_cadena_restaurante/iteracion1/src/main/java/org/iteracion1/Presentacion/IU_Jefe_de_Sala.java
@@ -1,0 +1,10 @@
+package Presentacion;
+
+public class IU_Jefe_de_Sala {
+
+	
+	public void mostrarMenuPrincipal() {
+		
+	}
+
+}
