--- src/main/java/ISOFT2_2021_BC01/ISOFT2_2021_BC01/Dominio/Estados.java
@@ -1,11 +1,0 @@
-package ISOFT2_2021_BC01.ISOFT2_2021_BC01.Dominio;
-
-public enum Estados { Libre, 
-	Reservada, 
-	Ocupada, 
-	pidiendo, 
-	espera_comida, 
-	servida, 
-	espera_cuenta, 
-	pagando, 
-	preparacion}
