--- src/main/java/ISOFT2_2021_BC01/ISOFT2_2021_BC01/Dominio/Jefe_de_Sala.java
@@ -1,24 +1,0 @@
-package ISOFT2_2021_BC01.ISOFT2_2021_BC01.Dominio;
-
-public class Jefe_de_Sala {
-
-    private int id_jefe_sala;
-    private String nombre;
-
-    public int getId_jefe_sala() {
-        return id_jefe_sala;
-    }
-
-    public void setId_jefe_sala(int id_jefe_sala) {
-        this.id_jefe_sala = id_jefe_sala;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-}
