--- src/main/java/ISOFT2_2021_BC01/ISOFT2_2021_BC01/Persistencia/BBDD.java
@@ -1,5 +1,0 @@
-package ISOFT2_2021_BC01.ISOFT2_2021_BC01.Persistencia;
-
-public class BBDD {
-	
-}
