--- src/main/java/ISOFT2_2021_BC01/ISOFT2_2021_BC01/Presentacion/IU_Jefe_de_Sala.java
@@ -1,10 +1,0 @@
-package ISOFT2_2021_BC01.ISOFT2_2021_BC01.Presentacion;
-
-public class IU_Jefe_de_Sala {
-
-	
-	public void mostrarMenuPrincipal() {
-		
-	}
-
-}
