--- src/test/java/ISOFT2_2021_BC01/ISOFT2_2021_BC01/AppTest.java
@@ -1,20 +1,0 @@
-package ISOFT2_2021_BC01.ISOFT2_2021_BC01;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
